requirements.txt
djangorestframework
django-debug-toolbar